Werror fix

+++ num/cc_num.h
@@ -13,7 +13,6 @@
 #ifndef CC_NUM_H
 # define CC_NUM_H
 
-# include <zconf.h>
 # undef numlen
 # undef numlen_base
 # undef atoi
@@ -21,11 +20,11 @@
 # undef putnbr
 # undef fdputnbr
 
-int					numlen(ssize_t num);
-int					numlen_base(ssize_t num, int base);
+int					numlen(int num);
+int					numlen_base(int num, int base);
 int					atoi(const char *str);
 long long			atoll(const char *str);
-ssize_t				putnbr(int n);
-ssize_t				fdputnbr(int n, int fd);
+int					putnbr(int n);
+int					fdputnbr(int n, int fd);
 
 #endif
